# Introduce vectorized parallel welford batchnorm for channels last path