Mainly added vectorization and optimized the workload, achieving a performance improvement of up to 1.169x in certain cases.